# WebContent/swagger/properties.json
{
  "swagger": "2.0",
  "info": {
    "title": "AltoroJ REST API",
    "version": "1.0.2"
  },
  "basePath": "/AltoroJ/api",
...
  "paths": {
...
    "/account": {
      "get": {
        "tags": [
          "2. Account"
        ],
        "operationId": "getAccount",
        "produces": [
          "application/json"
        ],
        "description": "Returns a list of all the accounts owned by the user",
        "parameters": [
         {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/account/{accountNo}": {
      "get": {
        "tags": [
          "2. Account"
        ],
        "operationId": "getAccountBalance",
        "produces": [
          "application/json"
        ],
        "description": "Returns details about a specific account",
        "parameters": [
         {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "description": "Account id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/account/{accountNo}/transactions": {
      "get": {
        "tags": [
          "2. Account"
        ],
        "operationId": "showLastTenTransactions",
        "description": "Returns the last 10 transactions attached to an account",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "description": "Account id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "2. Account"
        ],
        "description": "Return transactions between 2 specific dates",
        "operationId": "getTransactions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "name": "accountNo",
            "in": "path",
            "required": true,
            "description": "Account id",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "A start date and an end date",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dates"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad parameters: Please check provided values"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "501": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/admin/addUser": {
      "post": {
        "tags": [
          "5. Admin"
        ],
        "description": "Add new user",
        "operationId": "addUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Unauthorized request"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Error creating user"
          }
        },
        "parameters": [
        {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "in": "body",
            "name": "body",
            "description": "Details of a new user including first name, last name, desired username and a password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/newUser"
            }
          }
        ]
      }
    },
    "/admin/changePassword": {
      "post": {
        "tags": [
          "5. Admin"
        ],
        "description": "Change user password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "changePassword",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Unauthorized request"
          },
          "401": {
            "description": "Unauthorized request"
          },
          "500": {
            "description": "Error in changing the username password"
          }
        },
        "parameters": [
         {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Authorization token (provided upon successful login)",
            "type": "string"
         },
          {
            "in": "body",
            "name": "body",
            "description": "Information about the user password to be changed including id and new password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/changePassword"
            }
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "dates": {
      "type": "object",
      "required": [
        "startDate",
        "endDate"
      ],
      "properties": {
        "startDate": {
          "type": "string",
          "format": "date",
          "example": "2012-03-12"
        },
        "endDate": {
          "type": "string",
          "format": "date",
          "example": "2012-03-17"
        }
      }
    },
...
    "newUser": {
      "type": "object",
      "required": [
        "firstname",
        "lastname",
        "username",
        "password1",
        "password2"
      ],
      "properties": {
        "firstname": {
          "type": "string",
          "example": "Bilbo"
        },
        "lastname": {
          "type": "string",
          "example": "Baggins"
        },
        "username": {
          "type": "string",
          "example": "bilbob",
          "uniqueItems": true
        },
        "password1": {
          "type": "string",
          "format": "password",
          "example": "S3l3ctS0methingStr0ng5AsP@ssword"
        },
        "password2": {
          "type": "string",
          "format": "password",
          "example": "S3l3ctS0methingStr0ng5AsP@ssword"
        }
      }
    },
    "changePassword": {
      "type": "object",
      "required": [
        "username",
        "passwrod1",
        "password2"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "jdoe"
        },
        "password1": {
          "type": "string",
          "format": "password",
          "example": "Th1s!sz3nu3Passv0rd"
        },
        "password2": {
          "type": "string",
          "format": "password",
          "example": "Th1s!sz3nu3Passv0rd"
        }
      }
    }
  }
}